requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
WTForms==3.2.1
pillow==11.2.1
Flask-Mail==0.10.0